requirements.txt
colorama==0.4.1
Pillow==6.1.0
pypng==0.0.20
PyQRCode==1.2.1
qrcode==6.1
six==1.12.0
